// src/store/signIn/types.ts
export enum signInTypes {
  CREATE_ACCOUNT = "CREATE_ACCOUNT",
  CREATE_ACCOUNT_SUCCESS = "CREATE_ACCOUNT_SUCCESS",
  SET_LOADING = "SET_LOADING",
  OPEN_MODAL = "OPEN_MODAL",
  MAIL_MODAL = "MAIL_MODAL",
  RECOVER_PASSWORD = "RECOVER_PASSWORD",
  RECOVER_PASSWORD_SUCCESS = "RECOVER_PASSWORD_SUCCESS",
  SET_ERROR = "SET_ERROR",
  CLEAN_ERROR = "CLEAN_ERROR",
  LOGIN_ACCOUNT = "LOGIN_ACCOUNT",
  LOGIN_ACCOUNT_SUCCESS = "LOGIN_ACCOUNT_SUCCESS",
  REFRESH_TOKEN = "REFRESH_TOKEN",
  FORGOT_PASSWORD = "FORGOT_PASSWORD",
  REMOVE_LOGIN_STATE = "signInTypes/REMOVE_LOGIN_STATE",
}